% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdxmetroidentity-package.R
\docType{package}
\name{bdxmetroidentity-package}
\alias{bdxmetroidentity}
\alias{bdxmetroidentity-package}
\title{bdxmetroidentity: Bordeaux Metropole Graphic Design for Plots, R Markdown and Shiny
    Apps}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Produce ggplot2 plots, R markdown reports and shiny apps
    according to the graphic needs of the datalab of Bordeaux Metropole.
}
\details{
Ouvrir le pkgdown : \code{bdxmetroidentity::open_pkgdown()}
}
\author{
\strong{Maintainer}: Margot Brard \email{margot@thinkr.fr} (\href{https://orcid.org/0000-0001-6754-0659}{ORCID})

Authors:
\itemize{
  \item Cervan Girard \email{cervan@thinkr.fr} (\href{https://orcid.org/0000-0002-4816-4624}{ORCID})
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
