% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{change_theme}
\alias{change_theme}
\title{Switch beetween dark and light theme.}
\usage{
change_theme(input, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{input}{Result of input switch button}

\item{session}{Session}
}
\description{
Switch beetween dark and light theme.
}
