% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{create_palette_bdxmetro}
\alias{create_palette_bdxmetro}
\title{Return a fonction to interpolate a Bordeaux Metropole palette.}
\usage{
create_palette_bdxmetro(name_palette = "continuous", reverse = FALSE, ...)
}
\arguments{
\item{name_palette}{Character. Name of the palette. Should be one of "continous", "continous2", "discrete", "discrete2", "divergent" or "divergent2".}

\item{reverse}{Boolean. Indicates whether the palette should be reversed or not. Default is FALSE.}

\item{...}{Additional arguments to pass to colorRampPalette().}
}
\value{
A function to interpolate a Bordeaux Metropole palette.
}
\description{
Return a fonction to interpolate a Bordeaux Metropole palette.
}
\examples{
create_palette_bdxmetro("continuous")(6)
create_palette_bdxmetro("discrete")(6)
create_palette_bdxmetro("divergent")(6)
}
