% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{fetch_cookie}
\alias{fetch_cookie}
\title{Search for one cookie in the browser.}
\usage{
fetch_cookie(name, session = NULL)
}
\arguments{
\item{name}{Character. Name of the cookie to fetch.}

\item{session}{The `input` and `session` object from Shiny.}
}
\description{
Search for one cookie in the browser.
}
\details{
Taken from https://github.com/ColinFay/glouton — code under MIT License.
}
