% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{fluidpage_bdx}
\alias{fluidpage_bdx}
\title{Add external ressources and theme for fluidPage in shiny.}
\usage{
fluidpage_bdx(..., id = "dark_mode-page", input_id_button = "dark_mode")
}
\arguments{
\item{...}{Additional arguments of \code{shiny::\link{fluidPage}()}.}

\item{id}{Character. id of fluidPage.}

\item{input_id_button}{Character. id of the switch light/dark mode button.}
}
\value{
fluidpage
}
\description{
Add external ressources and theme for fluidPage in shiny.
}
