% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{html_document_bdxmetro}
\alias{html_document_bdxmetro}
\title{Convert a .Rmd file in .html format respecting the Bordeaux Metropole design.}
\usage{
html_document_bdxmetro(
  theme = "light",
  toc = TRUE,
  toc_depth = 2,
  toc_float = TRUE,
  code_folding = "hide",
  ...
)
}
\arguments{
\item{theme}{Character. Theme to be used. Either "light" (default) or "dark".}

\item{toc}{Boolean. To include a table of contents in the output. Default is TRUE.}

\item{toc_depth}{Numeric. Depth of headers to include in table of contents.}

\item{toc_float}{Boolean. To float the table of contents to the left of the document. Default is TRUE.}

\item{code_folding}{Character. Enable document readers to toggle the display of R code chunks.}

\item{...}{Additional arguments of \code{rmarkdown::\link{html_document}()}.}
}
\value{
An R Markdown object to be passed to
  \code{rmarkdown::\link{render}()}.
}
\description{
Convert a .Rmd file in .html format respecting the Bordeaux Metropole design.
}
