% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{html_vignette_bdxmetro}
\alias{html_vignette_bdxmetro}
\title{Convert a .Rmd file in .html format (vignette) respecting the Bordeaux Metropole design.}
\usage{
html_vignette_bdxmetro(
  theme = "light",
  toc = TRUE,
  toc_depth = 2,
  toc_float = TRUE,
  fig_width = 3,
  fig_height = 3,
  dev = "png",
  df_print = "default",
  keep_md = FALSE,
  readme = FALSE,
  self_contained = TRUE,
  ...
)
}
\arguments{
\item{theme}{Character. Theme to be used. Either "light" (default) or "dark".}

\item{toc}{Boolean. To include a table of contents in the output. Default is TRUE.}

\item{toc_depth}{Numeric. Depth of headers to include in table of contents.}

\item{toc_float}{Boolean. To float the table of contents to the left of the document. Default is TRUE.}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{df_print}{Method to be used for printing data frames. Valid values
include "default", "kable", "tibble", and "paged". The "default" method
uses a corresponding S3 method of \code{print}, typically
\code{print.data.frame}. The "kable" method uses the
\code{\link[knitr:kable]{knitr::kable}} function. The "tibble" method uses
the \pkg{tibble} package to print a summary of the data frame. The "paged"
method creates a paginated HTML table (note that this method is only valid
for formats that produce HTML). In addition to the named methods you can
also pass an arbitrary function to be used for printing data frames. You
can disable the \code{df_print} behavior entirely by setting the option
\code{rmarkdown.df_print} to \code{FALSE}. See
\href{https://bookdown.org/yihui/rmarkdown/html-document.html#data-frame-printing}{Data
frame printing section} in bookdown book for examples.}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{readme}{Use this vignette as the package README.md file (i.e. render
it as README.md to the package root). Note that if there are image files
within your vignette you should be sure to add README_files to .Rbuildignore}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}

\item{...}{Additional arguments passed to \code{\link[rmarkdown]{html_document}}. Please
note that \code{theme}, \code{fig_retina} and \code{highlight} are hard
coded. Setting any of those will yield an error.}
}
\value{
An R Markdown object to be passed to
  \code{rmarkdown::\link{render}()}.
}
\description{
Convert a .Rmd file in .html format (vignette) respecting the Bordeaux Metropole design.
}
