% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{init_cookie_theme}
\alias{init_cookie_theme}
\title{Init cookie for theme.}
\usage{
init_cookie_theme(input)
}
\arguments{
\item{input}{result of input switch button}
}
\value{
cookie inside browser
}
\description{
Init cookie for theme.
}
