% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{load_roboto_font}
\alias{load_roboto_font}
\title{Import Roboto font.}
\usage{
load_roboto_font()
}
\value{
Nothing.
}
\description{
Import Roboto font.
}
\examples{
library(ggplot2)

load_roboto_font()
extrafont::fonts() # available fonts

ggplot(data = iris) +
  aes(x = Sepal.Width, fill = Species) +
  geom_density(alpha = 0.6, color = NA) +
  labs(title = "Largeur des sépales des iris") +
  theme(
    title = element_text(family = "Roboto Light"),
    text = element_text(family = "Roboto")
    )
}
