% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{navbarpage_bdx}
\alias{navbarpage_bdx}
\title{Add external ressources and theme for navbarPage in shiny.}
\usage{
navbarpage_bdx(
  title,
  ...,
  id = "dark_mode-page",
  header = NULL,
  input_id_button = "dark_mode"
)
}
\arguments{
\item{title}{Character. The title to display in the navbar.}

\item{...}{Additional arguments of \code{shiny::\link{navbarPage}()}.}

\item{id}{Character. id of navbarPage}

\item{header}{Tag or list of tags to display as a common header above all tabPanels.}

\item{input_id_button}{Character. id of the switch light/dark mode button.}
}
\value{
navbarpage
}
\description{
Add external ressources and theme for navbarPage in shiny.
}
