% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picto.R
\name{read_picto}
\alias{read_picto}
\title{Import a Cairo SVG pictogram.}
\usage{
read_picto(name_picto)
}
\arguments{
\item{name_picto}{Character. Name of the SVG file without the .svg extension.}
}
\value{
A SVG object.
}
\description{
Import a Cairo SVG pictogram.
}
\examples{
car_picto <- read_picto(name_picto = "mag_car")
grImport2::grid.picture(picture = car_picto)
}
