% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scale_fill_bdxmetro_discrete}
\alias{scale_fill_bdxmetro_discrete}
\title{Discrete fill scale constructor for Bordeaux Metropole colors.}
\usage{
scale_fill_bdxmetro_discrete(name_palette = "discrete", reverse = FALSE, ...)
}
\arguments{
\item{name_palette}{Character. Name of the palette. Should be one of "discrete", "discrete2", "discrete3" (long).}

\item{reverse}{Boolean. Indicates whether the palette should be reversed or not. Default is FALSE.}

\item{...}{Additional arguments to pass to scale_fill_manual().}
}
\value{
A scale function.
}
\description{
Discrete fill scale constructor for Bordeaux Metropole colors.
}
\examples{
library(ggplot2)
ggplot(data = iris) +
  aes(x = Sepal.Width, fill = Species) +
  geom_density(alpha = 0.8) +
  scale_fill_bdxmetro_discrete()
}
