% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{show_palette_bdxmetro}
\alias{show_palette_bdxmetro}
\title{Show colors of a Bordeaux Metropole palette.}
\usage{
show_palette_bdxmetro(colors, show_codes = FALSE)
}
\arguments{
\item{colors}{Vector. Colors of the palette created with create_palette_bdxmetro().}

\item{show_codes}{Boolean. Indicates if hexadecimal codes should be printed or not. Default is FALSE.}
}
\value{
A ggplot2 plot.
}
\description{
Show colors of a Bordeaux Metropole palette.
}
\examples{
show_palette_bdxmetro(create_palette_bdxmetro("continuous")(6))
show_palette_bdxmetro(create_palette_bdxmetro("discrete")(6))
show_palette_bdxmetro(create_palette_bdxmetro("discrete")(6), show_codes = TRUE)
}
