% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{switch_button_theme}
\alias{switch_button_theme}
\title{Insert a switch button light/dark mode.}
\usage{
switch_button_theme(inputId, id = "dark_mode-page", page_type = "fluidpage")
}
\arguments{
\item{inputId}{Character. id of the input.}

\item{id}{Character. id of fluidPage or navbarPage.}

\item{page_type}{Character. Type of the page : "fluidpage" or "navbarpage".}
}
\value{
A fluidPage object.
}
\description{
Insert a switch button light/dark mode.
}
\details{
Inspired from https://rstudio.github.io/bslib/articles/theming.html.
}
