% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{theme_bdxmetro}
\alias{theme_bdxmetro}
\title{Light/Dark theme constructor for Bordeaux Metropole design.}
\usage{
theme_bdxmetro(theme = "light", ...)
}
\arguments{
\item{theme}{Character. Theme to be used. Either "light" (default) or "dark".}

\item{...}{Additional arguments of \code{bdxmetroidentity::\link{theme_bdxmetro_light}()} or \code{bdxmetroidentity::\link{theme_bdxmetro_dark}()}.}
}
\value{
A ggplot2 theme function.
}
\description{
Light/Dark theme constructor for Bordeaux Metropole design.
}
\examples{
library(ggplot2)
library(bdxmetroidentity)

ggplot(data = iris) +
  aes(x = Sepal.Width, fill = Species) +
  geom_density(alpha = 0.6, color = NA) +
  scale_fill_bdxmetro_discrete() +
  labs(title = "Largeur des sépales des iris",
       subtitle = "Ceci est un graphique qui respecte la charte Bordeaux Metropole",
       x = "Largeur des sépales",
       y = "Densité",
       fill = "Variété d'iris",
       caption = "Dark mode") +
  theme_bdxmetro(theme = "light")
}
