% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{theme_bdxmetro_dark}
\alias{theme_bdxmetro_dark}
\title{Dark theme constructor for Bordeaux Metropole design.}
\usage{
theme_bdxmetro_dark(
  regular_font_family = "Roboto",
  light_font_family = "Roboto Light",
  main_text_color = "white",
  secondary_text_color = "#9f9f9f",
  plot_title_family = light_font_family,
  plot_title_color = main_text_color,
  plot_title_size = 20,
  plot_subtitle_family = light_font_family,
  plot_subtitle_color = main_text_color,
  plot_subtitle_size = 12,
  axis_title_family = regular_font_family,
  axis_title_color = main_text_color,
  axis_title_size = 11,
  axis_text_family = regular_font_family,
  axis_text_color = secondary_text_color,
  axis_text_size = 9,
  legend_title_family = regular_font_family,
  legend_title_color = main_text_color,
  legend_title_size = 11,
  legend_text_family = regular_font_family,
  legend_text_color = main_text_color,
  legend_text_size = 9,
  legend_background_color = "#101010",
  legend_key_color = "#101010",
  facet_text_family = regular_font_family,
  facet_text_color = "white",
  facet_text_size = 12,
  facet_background = "#101010",
  facet_background_border = secondary_text_color,
  plot_caption_family = regular_font_family,
  plot_caption_color = secondary_text_color,
  plot_caption_size = 9,
  panel_background_color = "#101010",
  panel_grid_color = "#292929",
  plot_background_color = "#101010",
  plot_background_border_color = "#101010"
)
}
\arguments{
\item{regular_font_family}{Character. Regular font family. Default is Roboto.}

\item{light_font_family}{Character. Light font family. Default is Roboto Light.}

\item{main_text_color}{Character. Color for the main text color (i.e. the darkest).}

\item{secondary_text_color}{Character. Color for the seconday text color (i.e. the lightest).}

\item{plot_title_family}{Character. Font family for the plot title.}

\item{plot_title_color}{Character. Color for the plot title.}

\item{plot_title_size}{Numeric. Size for the plot title.}

\item{plot_subtitle_family}{Character. Font family for the plot subtitle.}

\item{plot_subtitle_color}{Character. Color for the plot subtitle.}

\item{plot_subtitle_size}{Numeric. Size for the plot subtitle.}

\item{axis_title_family}{Character. Font family for the axis titles (both x and y).}

\item{axis_title_color}{Character. Color for the the axis titles (both x and y).}

\item{axis_title_size}{Numeric. Size for the the axis titles (both x and y).}

\item{axis_text_family}{Character. Font family for the axis texts (both x and y).}

\item{axis_text_color}{Character. Color for the the axis texts (both x and y).}

\item{axis_text_size}{Numeric. Size for the the axis texts (both x and y).}

\item{legend_title_family}{Character. Font family for the legend title.}

\item{legend_title_color}{Character. Color for the legend title.}

\item{legend_title_size}{Numeric. Size for the legend title.}

\item{legend_text_family}{Character. Font family for the legend text.}

\item{legend_text_color}{Character. Color for the legend text.}

\item{legend_text_size}{Numeric. Size for the legend text.}

\item{legend_background_color}{Character. Color for the legend background.}

\item{legend_key_color}{Character. Color for the legend key.}

\item{facet_text_family}{Character. Font family for the facets texts.}

\item{facet_text_color}{Character. Color for the facets texts.}

\item{facet_text_size}{Numeric. Size for the facets texts.}

\item{facet_background}{Character. Color for the facets background.}

\item{facet_background_border}{Character. Color for the facets borders.}

\item{plot_caption_family}{Character. Font family for the caption.}

\item{plot_caption_color}{Character. Color for the caption.}

\item{plot_caption_size}{Numeric. Size for the caption.}

\item{panel_background_color}{Character. Color for the panel background.}

\item{panel_grid_color}{Character. Color for the panel grids.}

\item{plot_background_color}{Character. Color for the plot background.}

\item{plot_background_border_color}{Character. Color for the border of the plot background.}
}
\value{
A ggplot2 theme function.
}
\description{
Dark theme constructor for Bordeaux Metropole design.
}
\examples{
library(ggplot2)
library(bdxmetroidentity)

ggplot(data = iris) +
  aes(x = Sepal.Width, fill = Species) +
  geom_density(alpha = 0.6, color = NA) +
  scale_fill_bdxmetro_discrete() +
  labs(title = "Largeur des sépales des iris",
       subtitle = "Ceci est un graphique qui respecte la charte Bordeaux Metropole",
       x = "Largeur des sépales",
       y = "Densité",
       fill = "Variété d'iris",
       caption = "Dark mode") +
  theme_bdxmetro_dark()

ggplot(data = iris) +
  aes(x = Sepal.Width, fill = Species) +
  geom_density(alpha = 0.6, color = NA) +
  scale_fill_bdxmetro_discrete() +
  guides(fill = "none") +
  labs(title = "Largeur des sépales des iris",
       subtitle = "Ceci est un graphique qui respecte la charte Bordeaux Metropole",
       x = "Largeur des sépales",
       y = "Densité",
       fill = "Variété d'iris",
       caption = "Dark mode") +
  facet_grid(rows = vars(Species)) +
  theme_bdxmetro_dark()
}
