% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{theme_bdxmetro_dark_shiny}
\alias{theme_bdxmetro_dark_shiny}
\title{Create a dark bootstrap theme for Bordeaux Metropole shiny apps.}
\usage{
theme_bdxmetro_dark_shiny(
  bg = "#101010",
  fg = "#ffffff",
  primary = "#fd79da",
  secondary = "#039ed2",
  success = "#fd994e",
  info = "#fd5072",
  warning = "#5cadfd",
  danger = "#c4a5fd",
  base_font = font_google("Roboto"),
  heading_font = font_google("Roboto"),
  ...
)
}
\arguments{
\item{bg}{Character. Background color.}

\item{fg}{Character. Foreground color.}

\item{primary}{Character. Primary color.}

\item{secondary}{Character. Seconday color.}

\item{success}{Character. Success color.}

\item{info}{Character. Info color.}

\item{warning}{Character. Warning color.}

\item{danger}{Character. Danger color. Can be obtained with \code{bslib::\link{font_google}()}.}

\item{base_font}{Character. Base font. Can be obtained with \code{bslib::\link{font_google}()}.}

\item{heading_font}{Character. Heading font}

\item{...}{Additional arguments of \code{bslib::\link{bs_add_variables}()}.}
}
\value{
A bs_theme function.
}
\description{
Create a dark bootstrap theme for Bordeaux Metropole shiny apps.
}
