% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{use_glouton}
\alias{use_glouton}
\title{Add glouton to your shiny app.}
\usage{
use_glouton(online = TRUE)
}
\arguments{
\item{online}{Boolean. TRUE fetch the online version of js-cookie through
CDN. FALSE uses the file included in the package.}
}
\description{
Add glouton to your shiny app.
}
\details{
Taken from https://github.com/ColinFay/glouton — code under MIT License.
}
